add fitur category

+++ app/Http/Controllers/UserController.php
@@ -4,25 +4,46 @@ namespace App\Http\Controllers;
 
 use App\Models\Role;
 use App\Models\User;
+use App\Models\PermissionRole;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Auth;
 
 class UserController extends Controller
 {
     public function list()
     {
+        $PermissionRole = PermissionRole::getPermission('User', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
+        $data['PermissionAdd'] = PermissionRole::getPermission('Add User', Auth::user()->role_id);
+        $data['PermissionEdit'] = PermissionRole::getPermission('Edit User', Auth::user()->role_id);
+        $data['PermissionDelete'] = PermissionRole::getPermission('Delete User', Auth::user()->role_id);
+
         $data['getRecord'] = User::getRecord();
         return view('panel.user.list', $data);
     }
 
     public function add()
     {
+        $PermissionRole = PermissionRole::getPermission('Add User', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $data['getRole'] = Role::getRecord();
         return view('panel.user.add', $data);
     }
 
     public function insert(Request $request)
     {
+        $PermissionRole = PermissionRole::getPermission('Add User', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $request->validate([
             'email' => 'required|email|unique:users'
         ]);
@@ -39,6 +60,11 @@ class UserController extends Controller
 
     public function edit($id)
     {
+        $PermissionRole = PermissionRole::getPermission('Edit User', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $data['getRecord'] = User::getSingle($id);
         $data['getRole'] = Role::getRecord();
         return view('panel.user.edit', $data);
@@ -46,6 +72,11 @@ class UserController extends Controller
 
     public function update($id, Request $request)
     {
+        $PermissionRole = PermissionRole::getPermission('Edit User', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $user = User::getSingle($id);
         $user->name = trim($request->name);
         if (!empty($request->password)) {
@@ -59,6 +90,11 @@ class UserController extends Controller
 
     public function delete($id)
     {
+        $PermissionRole = PermissionRole::getPermission('Delete User', Auth::user()->role_id);
+        if (empty($PermissionRole)) {
+            abort(404);
+        }
+
         $user = User::getSingle($id);
         $user->delete();
